// src/interfaces/RawStudentRowValues.ts
export interface RawStudentRowValues {
  author: string;
  year: string;
  title: string;
  info: string;
  topic: string;
  discipline: string;
  theme: string;
  series0101: string;
  series0102: string;
  series0201: string;
  series0202: string;
  series0301: string;
  series0302: string;
  thumbnail: string;
  subtopic: string;
}